# inventarioBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Gestión de Inventario API",
        "description": "Sistema para gestionar inventario para empresas en crecimiento",
        "contact": {
            "name": "API Support",
            "email": "support@inventario.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/reportes/ventas": {
            "get": {
                "tags": [
                    "Reportes"
                ],
                "summary": "Generar reporte de ventas",
                "description": "Genera un reporte detallado de ventas con resumen general y detalle de productos vendidos",
                "operationId": "06d39f5dae6e89765174ec925e4f88ca",
                "parameters": [
                    {
                        "name": "fecha_inicio",
                        "in": "query",
                        "description": "Fecha de inicio (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fecha_fin",
                        "in": "query",
                        "description": "Fecha de fin (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reporte generado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "periodo": {
                                                    "type": "object"
                                                },
                                                "resumen": {
                                                    "properties": {
                                                        "numero_ventas": {
                                                            "type": "integer",
                                                            "example": 45
                                                        },
                                                        "ingresos_totales": {
                                                            "type": "number",
                                                            "example": 12500.5
                                                        },
                                                        "total_productos_vendidos": {
                                                            "type": "integer",
                                                            "example": 150
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "detalle_productos": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Fechas inválidas"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token",
                "name": "bearerAuth",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}